# feat(lint): add performance-related rules