--- a/packages/compiler-core/src/transforms/vBind.ts
+++ b/packages/compiler-core/src/transforms/vBind.ts
@@ -41,7 +41,7 @@ export const transformBind: DirectiveTransform = (dir, _node, context) => {
   }
 
   // same-name shorthand - :arg is expanded to :arg="arg"
-  if (!exp) {
+  if (exp === undefined) {
     if (arg.type !== NodeTypes.SIMPLE_EXPRESSION || !arg.isStatic) {
       // only simple expression is allowed for same-name shorthand
       context.onError(