… .github/workflows/autofix.yml
… package.json
… packages/compiler-sfc/src/style/preprocessors.ts
… packages/compiler-sfc/src/template/transformSrcset.ts
… packages/compiler-ssr/__tests__/utils.ts
… packages/compiler-ssr/src/ssrCodegenTransform.ts
… packages/compiler-ssr/src/transforms/ssrInjectCssVars.ts
… packages/compiler-ssr/src/transforms/ssrInjectFallthroughAttrs.ts
… packages/compiler-ssr/src/transforms/ssrTransformComponent.ts
… packages/compiler-ssr/src/transforms/ssrTransformElement.ts
… packages/compiler-ssr/src/transforms/ssrTransformTeleport.ts
… packages/compiler-ssr/src/transforms/ssrTransformTransition.ts
… packages/compiler-ssr/src/transforms/ssrVIf.ts
… packages/compiler-ssr/src/transforms/ssrVShow.ts
… packages/reactivity/src/computed.ts
… packages/reactivity/src/dep.ts
… packages/reactivity/src/effect.ts
… packages/reactivity/src/effectScope.ts
… packages/reactivity/src/watch.ts
… packages/runtime-core/src/apiAsyncComponent.ts
… packages/runtime-core/src/apiCreateApp.ts
… packages/runtime-core/src/apiInject.ts
… packages/runtime-core/src/apiWatch.ts
… packages/runtime-core/src/compat/global.ts
… packages/runtime-core/src/compat/instanceEventEmitter.ts
… packages/runtime-core/src/compat/renderFn.ts
… packages/runtime-core/src/compat/renderHelpers.ts
… packages/runtime-core/src/component.ts
… packages/runtime-core/src/componentEmits.ts
… packages/runtime-core/src/componentOptions.ts