Add 2 performance-related linting rules.

1. Numeric types should avoid `!== 0` comparisons.
2. Nullable object types should be compared using `=== undefined`/`!== undefined`/`=== null`/` !== null`.

#### Benchmark

`number !== 0`: 153.250917 ms
`!number`: 204.387375 ms
`object === undefined`: 184.250917 ms
`!object`: 358.422708 ms